+++ back-java/src/main/resources/db/migration/V03__Cricacao_De_Categoria_E_Mensagem.sql
@@ -1,13 +1,13 @@
 CREATE TABLE categoria
 (
-    id integer PRIMARY KEY auto_increment,
+    id bigint PRIMARY KEY auto_increment,
     nome varchar(200)
 );
 
 CREATE table mensagem
 (
-    id integer primary key auto_increment,
-    id_categoria integer,
+    id bigint primary key auto_increment,
+    id_categoria bigint,
     mensagem varchar(200),
     FOREIGN KEY (id_categoria) references categoria (id)
 );
