+++ back-java/src/main/resources/db/migration/V06__Criacao_log_msg.sql
@@ -1,6 +1,6 @@
 CREATE table log_torre
 (
-    id integer primary key auto_increment,
+    id bigint primary key auto_increment,
     id_vant varchar(255) not null,
     mensagem varchar(255) not null,
     data_envio timestamp not null
